fix: Strict type mismatch pada authorizeEvent user_id comparison
Ganti !== ke (int) cast agar perbandingan user_id tidak gagal
akibat perbedaan tipe string/int antara DB dan auth()->id().

+++ app/Http/Controllers/User/EventController.php
@@ -374,7 +374,7 @@ class EventController extends Controller
             return;
         }
 
-        if ($event->user_id !== auth()->id()) {
+        if ((int) $event->user_id !== (int) auth()->id()) {
             abort(403, 'Anda tidak memiliki akses ke undangan ini.');
         }
     }
